Bug fixing, changes in POST JSON handling for api, fully implemented reCAPTCHA for create-box API, changed password hashing.
Added reCAPTCHA secret key variable to .env.example. Fully implemented and tested reCAPTCHA feature for create-box api.
Updated password hashing in create-box api from hmac to hash for consistency with verifyPasswordFromHeader funciton.
Updated the post api so that they do not depend on the ordering of the contents of the post json. Fixed post api bugs when handling undefined
variables in JSON.